requirements.txt
RPi.GPIO==0.5.11
pyOpenSSL==0.15.1
pyalsaaudio==0.8.2
requests==2.7.0
netifaces==0.10.4
scapy==2.3.1
python-nmap==0.4.1
tinydb
netaddr
bottle
